+++ game2020/game2020/World/Level1.cs
@@ -1,15 +1,19 @@
-﻿using Game1;
+﻿using game2020.Interfaces;
+using game2020.World;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using RefactoringCol;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace RefactoringCol
+namespace Game1
 {
-    class Level1 : Levels
+    class Level1: Levels
     {
         public Level1()
         {
-            Generate(new int[,]
+            GenerateLevel(new int[,]
             {
                 { 1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,},
                 { 2,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,},
@@ -20,6 +24,23 @@ namespace RefactoringCol
                 { 2,0,0,0,0,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,},
                 { 2,1,1,1,1,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,2,}
             }, 64);
+        }
+
+        public void GenerateLevel(int[,] Map, int size)
+        {
+            for (int x = 0; x < Map.GetLength(1); x++)
+            {
+                for (int y = 0; y < Map.GetLength(0); y++)
+                {
+                    int number = Map[y, x];
+
+                    if (number > 0)
+                        CollisionTiles.Add(new CollisionTiles(number, new Rectangle(x * size, y * size, size, size)));
+
+                    width = (x + 1) * size;
+                    height = (y + 1) * size;
+                }
+            }
         }
     }
 }
